verify: Refactor the verify package to have common code
The verify package should be usable from multiple clients. Refactor
it to separate dependencies on the payload in preparation for a move
to an external library.
